lessons/05-data-retrieval/code/server/core/config.py
# server/core/config.py
import os
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

class Settings:
    OPENAI_API_KEY: str = os.getenv("OPENAI_API_KEY")
    ALLOWED_ORIGINS: list = ["http://localhost:3000"]

settings = Settings()
